Week_1_and_2/Resume_Analyser/requirements.txt
# Core (no FastAPI)
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain
langchain>=0.1.0
langchain-openai>=0.0.5

# PDF Processing
pypdf>=3.17.0
python-docx>=1.1.0

# Embeddings & Vector Store
numpy>=1.24.0
sentence-transformers>=2.2.0
qdrant-client>=1.7.0

# Utilities
python-dotenv>=1.0.0
